backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
motor==3.6.0
pymongo==4.9.1
pydantic-settings==2.5.2
numpy==1.26.4
Pillow==10.4.0
tensorflow==2.17.0
requests==2.32.3
pyserial==3.5
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
